Call: Low the video watchdog timeout to 1 second
Switch back to the avatar if there was at least one second of not receiving any
new video frames to display.

With this change the reaction time will be between 1 (best case) and 2
(worse case) seconds from the point the last frame went into the sync.
Before it was 5 seconds (best case) and up to 10 seconds (worst case),
which doesn't make for a very good experience