individual_menu_add_personas: add the block menu item, if needed
This allow user to select the personas he wants to block.

Note that I didn't pass an EmpathyContact as second argument to
empathy_individiual_block_menu_item_new() as we currently do with most
item_new() methods. Instead I just craft a new individual containing only the
persona we care about. This approach is much cleaner as we don't have to deal
with 2 different code paths. I filed bug #666163 to do the same with the
existing functions.

https://bugzilla.gnome.org/show_bug.cgi?id=665038