Use a localized format when displaying times
Unfortunatelly there is no localized format displaying only the hour and
minutes (bug #668323) so we'll have to use %X for now which also display the
seconds.

EMPATHY_DATE_FORMAT_DISPLAY_SHORT was already localized so this one is fine.

That also means we don't have to translate those formats any more as we'll
just rely on the system env.

https://bugzilla.gnome.org/show_bug.cgi?id=608474