Don't unhighlight chat tabs when more messages are received
Travis noticed that if you receive a message in which you are not
mentioned after one in which you are, the highlight on the chat tab is
cleared. This turned out to be because the text of the chat tab label
was set in two different places in empathy-chat-window.c: one which
takes should_highlight into account, and one which does not.

This patch makes EmpathyChat keep track of whether it should be
highlighted (reusing empathy_chat_messages_read(), which is called to
acknowledge messages as having been seen by the user, to clear the
flag), and then uses that from one of the label-updating code paths and
deletes the other.

https://bugzilla.gnome.org/show_bug.cgi?id=669823