call-window: use a ClutterBoxLayout to separate previews/toolbar
It's way easier to use a ClutterBoxLayout to ensure the preview
rectangles and the floating toolbar don't overlap.
This way, we can get rid of some complicated manual UI layouting code
and just a margin to set the bottom spacing for the floating toolbar.
Also, this makes the code not dependent on a hardcoded toolbar size
anymore.

https://bugzilla.gnome.org/show_bug.cgi?id=672560